refactor: moved output class and pushing to BN to common, as may need to send results outside of container on job fail
