+++ grading-vm/grading_container/build_script/json_helpers/grading_script_command.py
@@ -1,4 +1,4 @@
-from grading_container.validations.grading_job_json_types import GradingScriptCommandJSON
+from common.types.grading_job_json_types import GradingScriptCommandJSON
 
 def is_conditional_command(cmd: GradingScriptCommandJSON):
   return "on_false" in cmd and "on_true" in cmd and "check" in cmd
