+++ grading-vm/grading_container/grading_script/bash_grading_script_command.py
@@ -1,8 +1,8 @@
 from subprocess import CalledProcessError, CompletedProcess, run, TimeoutExpired
 from typing import List
+from common.job_output.grading_job_output import GradingJobOutput
 from grading_container.grading_script.grading_script_command import GradingScriptCommand
 from grading_container.grading_script.grading_script_command_response import GradingScriptCommandResponse
-from grading_container.output.grading_job_output import GradingJobOutput
 
 
 class BashGradingScriptCommand:
